# refactor: remove Arc usage in cache handle