--- a/foyer-common/src/lib.rs
+++ b/foyer-common/src/lib.rs
@@ -14,7 +14,6 @@
 
 #![cfg_attr(coverage_nightly, feature(coverage_attribute))]
 
-pub mod arc_key_hash_map;
 pub mod async_batch_pipeline;
 pub mod async_queue;
 pub mod asyncify;