… foyer-common/src/arc_key_hash_map.rs
… foyer/src/hybrid/cache.rs