@MrCroxx 2024-05-15
## Benchmark

### Command

```bash
rm -rf /p44pro/foyer && cargo build --release && RUST_BACKTRACE=1 RUST_LOG=warn ./target/release/foyer-bench --dir /p44pro/foyer --mem 64 --disk 102400 --file-size 64 --get-range 10000 --flushers 12 --reclaimers 12 --time 60 --writers 64 --w-rate 8 --admission-rate-limit 500 --readers 1024 --r-rate 1 --runtime --compression none --distribution zipf --distribution-zipf-s 0.8
```

### main (https://github.com/MrCroxx/foyer/pull/475#issuecomment-2111655699)

```
Total:
disk total iops: 6775.8
disk total throughput: 467.8 MiB/s
disk read iops: 131.9
disk read throughput: 8.8 MiB/s
disk write iops: 6643.9
disk write throughput: 459.0 MiB/s
insert iops: 6706.9/s
insert throughput: 419.2 MiB/s
insert lat p50: 3us
insert lat p90: 5us
insert lat p99: 8us
insert lat p999: 65us
insert lat p9999: 146us
insert lat p99999: 211us
insert lat pmax: 291us
get iops: 15848.1/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 114us
get hit lat p90: 347us
get hit lat p99: 607us
get hit lat p999: 5151us
get hit lat p9999: 6943us
get hit lat p99999: 7839us
get hit lat pmax: 7871us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```

### PR

```
Total:
disk total iops: 6771.5
disk total throughput: 468.4 MiB/s
disk read iops: 128.5
disk read throughput: 8.5 MiB/s
disk write iops: 6642.9
disk write throughput: 459.9 MiB/s
insert iops: 6720.5/s
insert throughput: 420.1 MiB/s
insert lat p50: 2us
insert lat p90: 4us
insert lat p99: 8us
insert lat p999: 69us
insert lat p9999: 142us
insert lat p99999: 200us
insert lat pmax: 259us
get iops: 15848.6/s
get miss: 0.00%
get throughput: 1.0 GiB/s
get hit lat p50: 48us
get hit lat p90: 323us
get hit lat p99: 559us
get hit lat p999: 1655us
get hit lat p9999: 2111us
get hit lat p99999: 6335us
get hit lat pmax: 6559us
get miss lat p50: 0us
get miss lat p90: 0us
get miss lat p99: 0us
get miss lat p999: 0us
get miss lat p9999: 0us
get miss lat p99999: 0us
get miss lat pmax: 0us
```